feat: migration users to ts

+++ src/resources/users/user.router.ts
@@ -1,8 +1,10 @@
-const router = require('express').Router();
-const User = require('./user.model');
-const usersService = require('./user.service');
+import { Request, Response, Router } from 'express';
+import usersService from './user.service';
+import User from './user.model';
 
-router.route('/').get(async (req, res) => {
+const router = Router();
+
+router.route('/').get(async (_req: Request, res: Response) => {
   try {
     const users = await usersService.getAll();
     res.json(users.map(User.toResponse));
@@ -11,17 +13,20 @@ router.route('/').get(async (req, res) => {
   }
 });
 
-router.route('/:userId').get(async (req, res) => {
+router.route('/:userId').get(async (req: Request, res: Response) => {
   try {
-    const { userId } = req.params
-    const user = await usersService.getById(userId);
-    res.json(User.toResponse(user));
+    const { userId } = req.params;
+    if (typeof userId === 'string') {
+      const user = await usersService.getById(userId);
+      res.json(User.toResponse(user));
+    }
+
   } catch ({ message }) {
     res.status(404).send(message);
   }
 });
 
-router.route('/').post(async (req, res) => {
+router.route('/').post(async (req: Request, res: Response) => {
   try {
     const user = await usersService.create(
       new User({ ...req.body })
@@ -32,21 +37,27 @@ router.route('/').post(async (req, res) => {
   }
 });
 
-router.route('/:userId').put(async (req, res) => {
+router.route('/:userId').put(async (req: Request, res: Response) => {
   try {
     const { userId } = req.params;
-    const user = await usersService.update(userId, req.body);
-    res.json(User.toResponse(user));
+    if (typeof userId === 'string') {
+      const user = await usersService.update(userId, req.body);
+      res.json(User.toResponse(user));
+    }
+
   } catch ({ message }) {
     res.status(404).send(message);
   }
 });
 
-router.route('/:userId').delete(async (req, res) => {
+router.route('/:userId').delete(async (req: Request, res: Response) => {
   try {
     const { userId } = req.params;
-    await usersService.remove(userId);
-    res.status(204).json(`user is deleted with id = ${userId}`);
+    if (typeof userId === 'string') {
+      await usersService.remove(userId);
+      res.status(204).json(`user is deleted with id = ${userId}`);
+    }
+
   } catch ({ message }) {
     res.status(404).send(message);
   }
